Fix in bbox clipping
